Reorganize code base
Repackage classes for Person and Tags.
